requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
argparse==1.4.0
    # via -r requirements.in
build==0.10.0
    # via pip-tools
click==8.1.3
    # via pip-tools
packaging==23.0
    # via build
pip-tools==6.12.3
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python3-dtls==1.3.0
    # via -r requirements.in
tomli==2.0.1
    # via build
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
